backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.0
Pillow==11.2.1
drf-spectacular==0.28.0
django-filter==25.1
psycopg2-binary==2.9.10
